Remove dependencies for kuttl targets
To avoid interfering with dev workflows and better align the kuttl
targets  with the rest of the targets, this commit removes the
dependencies of the kuttl targets so it does not affect the present
state of the cluster and assumes the user will run beforehand the
necessary commands to set everything up.

There will be two kuttl targets for each operator, an operator_kuttl
target with all the dependencies and clenaup and a new
operator_kuttl_run which simply runs the kuttl tests.